requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

celery==5.4.0
eventlet==0.38.2
redis==5.2.1

channels==4.1.0
channels-redis==4.2.0
websockets==14.1

drf-yasg==1.21.8

psycopg2-binary==2.9.10

python-dotenv==1.0.1

pytest==8.3.4
pytest-asyncio==0.25.0


requests==2.32.3
